packages reordering: moving query and queryable into query package
