+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/query/Query.kt
@@ -12,7 +12,7 @@
 //   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 //
 
-package io.zenoh.queryable
+package io.zenoh.query
 
 import io.zenoh.ZenohType
 import io.zenoh.selector.Selector
