+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/GetTest.kt
@@ -19,7 +19,7 @@ import io.zenoh.sample.SampleKind
 import io.zenoh.bytes.into
 import io.zenoh.config.Config
 import io.zenoh.query.Reply
-import io.zenoh.queryable.Queryable
+import io.zenoh.query.Queryable
 import io.zenoh.selector.Parameters
 import io.zenoh.selector.Selector
 import io.zenoh.selector.intoSelector
